requirements.txt
absl-py==2.2.2
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.58.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
keras==3.9.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.9
narwhals==1.40.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.15.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
protobuf==4.25.7
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.0
six==1.17.0
smmap==5.0.2
streamlit==1.45.1
streamlit-drawable-canvas==0.9.3
tenacity==9.1.2
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
toml==0.10.2
tornado==6.5
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
